make patch feasible
